# database/sql: Replace ctx.Done() with ctx.Err()