Prior to this commit, we used ctx.Done() to test for cancellation
without multiplexing over other channels like:

  select {
    default:
    case <-ctx.Done():
      // handle cancellation
  }

context.Context.Done() lazily allocates the internal channel used and
there is some overhead to entering a select block.

This commit replaces lone ctx.Done() calls with ctx.Err() like below:

  if err := ctx.Err(); err != nil {
    // handle cancellation
  }

This potentially avoids the channel allocation, avoids the overhead of
unnecessarily entering a select block, and saves a few lines and indent
levels.  Still, it is probably more correct to use ctx.Err() when not
multiplexing over other channels.

benchstat output shows a measurable but negligible performance
improvement and no change to memory footprint.

name                     old time/op    new time/op    delta
ConcurrentDBExec-8         6.28ms ± 0%    6.27ms ± 0%   ~     (p=1.000 n=1+1)
ConcurrentStmtQuery-8      5.61ms ± 0%    5.52ms ± 0%   ~     (p=1.000 n=1+1)
ConcurrentStmtExec-8       4.29ms ± 0%    4.28ms ± 0%   ~     (p=1.000 n=1+1)
ConcurrentTxQuery-8        5.55ms ± 0%    5.57ms ± 0%   ~     (p=1.000 n=1+1)
ConcurrentTxExec-8         1.40ms ± 0%    1.39ms ± 0%   ~     (p=1.000 n=1+1)
ConcurrentTxStmtQuery-8    1.63ms ± 0%    1.63ms ± 0%   ~     (p=1.000 n=1+1)
ConcurrentTxStmtExec-8      454µs ± 0%     447µs ± 0%   ~     (p=1.000 n=1+1)
ConcurrentRandom-8         4.66ms ± 0%    4.63ms ± 0%   ~     (p=1.000 n=1+1)
ManyConcurrentQueries-8    10.4µs ± 0%    10.3µs ± 0%   ~     (p=1.000 n=1+1)

name                     old alloc/op   new alloc/op   delta
ConcurrentDBExec-8         2.49MB ± 0%    2.50MB ± 0%   ~     (p=1.000 n=1+1)
ConcurrentStmtQuery-8      2.25MB ± 0%    2.25MB ± 0%   ~     (p=1.000 n=1+1)
ConcurrentStmtExec-8       1.71MB ± 0%    1.71MB ± 0%   ~     (p=1.000 n=1+1)
ConcurrentTxQuery-8        2.37MB ± 0%    2.44MB ± 0%   ~     (p=1.000 n=1+1)
ConcurrentTxExec-8          587kB ± 0%     587kB ± 0%   ~     (all equal)
ConcurrentTxStmtQuery-8     745kB ± 0%     745kB ± 0%   ~     (p=1.000 n=1+1)
ConcurrentTxStmtExec-8      108kB ± 0%     108kB ± 0%   ~     (p=1.000 n=1+1)
ConcurrentRandom-8         1.63MB ± 0%    1.66MB ± 0%   ~     (p=1.000 n=1+1)
ManyConcurrentQueries-8    3.76kB ± 0%    3.76kB ± 0%   ~     (all equal)

name                     old allocs/op  new allocs/op  delta
ConcurrentDBExec-8          14.6k ± 0%     14.6k ± 0%   ~     (all equal)
ConcurrentStmtQuery-8       17.1k ± 0%     17.1k ± 0%   ~     (all equal)
ConcurrentStmtExec-8        3.17k ± 0%     3.17k ± 0%   ~     (all equal)
ConcurrentTxQuery-8         19.6k ± 0%     19.6k ± 0%   ~     (all equal)
ConcurrentTxExec-8          14.1k ± 0%     14.1k ± 0%   ~     (all equal)
ConcurrentTxStmtQuery-8     19.2k ± 0%     19.2k ± 0%   ~     (all equal)
ConcurrentTxStmtExec-8      2.17k ± 0%     2.17k ± 0%   ~     (all equal)
ConcurrentRandom-8          13.9k ± 0%     13.9k ± 0%   ~     (p=1.000 n=1+1)
ManyConcurrentQueries-8      17.0 ± 0%      17.0 ± 0%   ~     (all equal)

This PR will be imported into Gerrit with the title and first
comment (this text) used to generate the subject and body of
the Gerrit change.

**Please ensure you adhere to every item in this list.**

More info can be found at https://github.com/golang/go/wiki/CommitMessage

+ The PR title is formatted as follows: `net/http: frob the quux before blarfing`
  + The package name goes before the colon
  + The part after the colon uses the verb tense + phrase that completes the blank in,
    "This change modifies Go to ___________"
  + Lowercase verb after the colon
  + No trailing period
  + Keep the title as short as possible. ideally under 76 characters or shorter
+ No Markdown
+ The first PR comment (this one) is wrapped at 76 characters, unless it's
  really needed (ASCII art, table, or long link)
+ If there is a corresponding issue, add either `Fixes #1234` or `Updates #1234`
  (the latter if this is not a complete fix) to this comment
+ If referring to a repo other than `golang/go` you can use the
  `owner/repo#issue_number` syntax: `Fixes golang/tools#1234`
+ We do not use Signed-off-by lines in Go. Please don't add them.
  Our Gerrit server & GitHub bots enforce CLA compliance instead.
+ Delete these instructions once you have read and applied them